missedGuesses not working.
I'm new to python and just started learning it in my intro to programming class. We have to create a hangman game. I have the code written but I can't seem to figure out why the missed guesses isn't working. It lists the correct guesses in the missed section. Any help would be grateful. We are using automate the boring stuff with python by Al Sweigart. Chapter 8 even gives the code and I'm not seeing the difference.  Thank you. - Rachel